apis/image/src/main/java/com/comcast/cats/image/ImageUtil.java
/**
 * Copyright 2014 Comcast Cable Communications Management, LLC
 *
 * This file is part of CATS.
 *
 * CATS is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * CATS is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with CATS.  If not, see <http://www.gnu.org/licenses/>.
 */
package com.comcast.cats.image;

import java.awt.image.BufferedImage;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import javax.imageio.ImageIO;

/**
 * Class contains Imaging utility functions.
 * @author mzmuda
 */
public class ImageUtil {
    /**
     * Converts a BufferedImage into a byte[].
     *
     * @param img The BufferedImage instance.
     * @return The BufferedImage as a byte[].
     * @throws IOException if there is a issue encoding the jpeg.
     */
    public static byte[] bufferedImageToByteArray(BufferedImage img) throws IOException {
        ByteArrayOutputStream os = new ByteArrayOutputStream();
        ImageIO.write(img, "JPEG", os);
       
        return os.toByteArray();
    }
}
